🛠 #358 iOS와 shake 흔들기 로직 맞추기

+++ feature/danggn/src/main/java/com/mashup/feature/danggn/data/danggn/DanggnGameController.kt
@@ -34,7 +34,7 @@ class DanggnGameController @Inject constructor(
     private var onShakeListener: (() -> Unit)? = null
 
     fun start(
-        threshold: Int,
+        threshold: Float,
         interval: Long,
         goldenDanggnPercent: Int
     ) {
